Replaced 4x4 grid with 8x8 grid. Removed obsolete ForderAITest file.

+++ forder/src/main/java/com/example/game/PatternEngine.java
@@ -11,8 +11,7 @@ public final class PatternEngine {
             Cell[][] cells,
             Direction direction
     ) {
-
-        int board = ForderBoardConverter.encode(cells);
+        long board = ForderBoardConverter.encode(cells);
 
         Player player = direction == Direction.RIGHT
                 ? Player.RED
